emotion_music/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
django-autoslug==1.9.9  
idna==3.11
pillow==12.1.1
requests==2.32.5        
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
gunicorn
whitenoise
dj-database-url
python-dotenv
psycopg2-binary
